move files from cmd to pkg/cmd
