+++ pkg/cmd/root.go
@@ -5,7 +5,6 @@ import (
 
 	"github.com/gogo/protobuf/types"
 	"github.com/solo-io/gloo/pkg/api/types/v1"
-	"github.com/solo-io/gloo/pkg/storage"
 	//	"github.com/pkg/errors"
 	"time"
 
@@ -16,25 +15,6 @@ import (
 const (
 	allOrigins = "all-origins"
 )
-
-type GlooClient struct {
-	Store storage.Interface
-}
-
-type RouteList struct {
-	// name of the origin service where the routes will be applied
-	// leave empty to apply to all origins
-	OriginServiceName string `json:"origin_service_name"`
-	// name of the upstream service for all the routes
-	// leave empty to apply to all destinations
-	DestinationServiceName string `json:"destination_service_name"`
-}
-
-type Route struct {
-	Matcher *v1.RequestMatcher `json:"matcher"`
-	Config  *types.Struct      `json:"config"`
-	// Destination is implicit
-}
 
 func (c *GlooClient) ConfigureService(serviceType string, retries uint32) error {
 	fmt.Println("MK configing service")
